% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logOfGamma.R
\name{gammaln}
\alias{gammaln}
\title{Computes the natural logarithm of the gamma function.}
\usage{
gammaln(x)
}
\arguments{
\item{x}{A numeric vector of positive numbers.}
}
\description{
For values larger than 12, an approximation is used.
}
\examples{
gammaln(5)
gammaln(50)
}

