% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logOfGamma.R
\name{gammaln_internal}
\alias{gammaln_internal}
\title{Computes the natural logarithm of the gamma function for values larger than 12.}
\usage{
gammaln_internal(x)
}
\arguments{
\item{x}{A numeric value of length 1 greater than 12}
}
\description{
Uses the approximation in Hart et al, Computer Approximations 1968.
}
\examples{
gammaln_internal(50)
}

